moving tests with xor to test_shelltask_inputspec.py
